refactor some code into components
